+++ crates/game/assets/src/codec/attack.rs
@@ -10,7 +10,7 @@ use serde::{Deserialize, Serialize};
 pub struct AttackCodec {
     pub format: u8,
     pub duration: u64,
-    pub stamina_cost: usize,
+    pub stamina_cost: u32,
     pub animation: ResourceLocation<AnimationResource>,
     pub particle_sprite: ResourceLocation<CharacterSpriteResource>,
     pub key_frames: Vec<KeyFrameCodec>,
